% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domin.r
\name{domin}
\alias{domin}
\title{User-Definable Dominance Analysis}
\usage{
domin(
  formula_overall,
  R_regression,
  fitstat_function,
  sets = NULL,
  all = NULL,
  ...
)
}
\arguments{
\item{formula_overall}{R formula for use in R_regression.}

\item{R_regression}{The R regression model called.}

\item{fitstat_function}{The fit statistic used for dominance analysis.}

\item{sets}{List of vectors of names; each list element's vector is a set.}

\item{all}{Vector of names in all subsets.}
}
\description{
Dominance analysis procedure accepting user-defined models and fit statistics.  An R port of Stata's -domin- module.
}
\examples{
domin()
}
\keyword{importance}
\keyword{relative}
