% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domin_scalar.r
\name{domin_scalar}
\alias{domin_scalar}
\title{Internal dominance analysis computation function assuming a scalar
input}
\usage{
domin_scalar(component_list, conditional, complete, reverse)
}
\description{
Internal dominance analysis computation function assuming a scalar
input
}
\keyword{internal}
