% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domir_scalar.r
\name{domir_scalar}
\alias{domir_scalar}
\title{Internal dominance analysis computation function assuming scalar
output}
\usage{
domir_scalar(
  fitting_fun,
  args_list,
  cons_args,
  full_fit,
  conditional,
  complete,
  reverse
)
}
\description{
Internal dominance analysis computation function assuming scalar
output
}
\keyword{internal}
