% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domme.r
\name{domme}
\alias{domme}
\title{Internal dominance analysis computation function}
\usage{
domme(component_list, conditional, complete, reverse)
}
\description{
Internal dominance analysis computation function
}
\keyword{internal}
