% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list.R
\name{formula_list}
\alias{formula_list}
\title{A \code{\link{list}} composed of \code{formulas}}
\usage{
formula_list(...)
}
\arguments{
\item{...}{\code{formula}s, possibly named}
}
\value{
A \code{list} of class \code{formula_list}.
}
\description{
Defines a list object composed of \code{formula}s. The purpose of this
class of object is to impose structure of the list to ensure that it
can be used to obtain response-term pairs and will be able to be
parsed in \code{\link{domir}}.
}
\details{
The \code{formula_list} requires that each element of the list is a \code{formula}
and that each \code{formula} is unique with a different, non-\code{NULL}
dependent variable/response.
}
