% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domir.r
\name{formula_parse}
\alias{formula_parse}
\title{Internal formula parsing function}
\usage{
formula_parse(.obj)
}
\description{
Internal formula parsing function to facilitate re-construction
of a formula using \code{reformulate()}

Not intended to be called by the user.
}
\keyword{internal}
