% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_dmn.r
\name{metric_dmn}
\alias{metric_dmn}
\title{Control function for \code{metric} argument of \code{domin}}
\usage{
metric_dmn(.fct, .flt_fct, ...)
}
\arguments{
\item{.fct}{Fit statistic/metric producing function; pre-defined or
anonymous.}

\item{.flt_fct}{Filter function; pre-defined or anonymous.}

\item{...}{Passes arguments to the fit statistic/metric producing function
in \code{.fct}.}
}
\value{
Returns the matched call/itself as a list representing the
elements of the unevaluated function call.
}
\description{
Supplies the fit statistic/metric producing function,
function to filter the result of the fit metric producing function, and
any additional arguments to the fit statistic/metric producing function.
}
\details{
\code{domin} has three S3 methods that use either
\link{`formula`}, \link{`Formula`}, or \link{`list`} objects
}
