% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dmn.r
\name{model_dmn}
\alias{model_dmn}
\title{Control function for \code{model} argument of \code{domin}}
\usage{
model_dmn(.fct, .lnk_fct = NULL, ...)
}
\arguments{
\item{.fct}{Modeling function; pre-defined or anonymous.}

\item{.lnk_fct}{Linking function; pre-defined or anonymous.}

\item{...}{Passes arguments to the modeling function in \code{.fct}.}
}
\value{
Returns the matched call/itself as a list representing the
elements of the unevaluated function call.
}
\description{
Supplies the model to be dominance analyzed, optional
linking functions, and all arguments to the model.
}
\details{
\code{domin} has three S3 methods that use either
\link{`formula`}, \link{`Formula`}, or \link{`list`} objects
}
