% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_control.r
\name{reg_control}
\alias{reg_control}
\title{Control function for \code{reg} argument of \code{domin}}
\usage{
reg_control(.fct, .lnk_fct, ...)
}
\arguments{
\item{.fct}{Modeling function; pre-defined or anonymous}

\item{.lnk_fct}{Linking function; pre-defined or anonymous}

\item{...}{Passes arguments to the function in \code{.fct}.}
}
\value{
boo
}
\description{
Controls the relative importance algorithm applied in the \code{tweed} function.
}
\details{
\code{domin} has three S3 methods that use either \link{\code{formula}}, \link{\code{Formula}}, or \link{\code{list}} objects
}
\examples{
hi
}
