% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domin.r
\name{summary.domin}
\alias{summary.domin}
\title{Summary method for \code{domin}}
\usage{
\method{summary}{domin}(object, ...)
}
\arguments{
\item{object}{an object of class "domin".}

\item{...}{further arguments passed to or from other methods. Not used currently.}
}
\value{
The originally submitted "domin" object with an additional \code{Strongest_Dominance} element added.
\describe{
\item{\code{Strongest_Dominance}}{Matrix comparing the independent variable in the first row to the independent variable in the third row.  The second row denotes the strongest designation between the two independent variables.}
}
}
\description{
Reports dominance designation results from the \code{domin} class object.
}
\details{
The summary method for class \code{domin} is used for obtaining the strongest dominance designations (i.e., general, conditional, or complete) among the independent variables.
}
