% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domir.r
\name{summary.domir}
\alias{summary.domir}
\title{Summary method for \code{domir}}
\usage{
\method{summary}{domir}(object, ...)
}
\arguments{
\item{object}{an object of class "domir".}

\item{...}{further arguments passed to or from other methods.
Not used currently.}
}
\value{
The submitted "domir" object with an additional
\code{Strongest_Dominance} element added.
\describe{
\item{\code{Strongest_Dominance}}{Matrix comparing the element in the first
row to the element in the third row.  The second row denotes the strongest
designation between the two names.}
}
}
\description{
Reports dominance designation results from the \code{domir}
class object.
}
\details{
The summary method for class \code{domir} objects is used for obtaining
the strongest dominance designations (i.e., general, conditional, or
complete) among all pairs of dominance analyzed names.
}
