% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domin.r
\name{domin}
\alias{domin}
\title{User-Definable Dominance Analysis}
\usage{
domin(
  formula_overall,
  reg,
  fitstat,
  sets = NULL,
  all = NULL,
  complete = TRUE,
  ...
)
}
\arguments{
\item{formula_overall}{Formula for use in reg.}

\item{reg}{The function implementing the regression model called.}

\item{fitstat}{List indicating function implemeting fit statistic and list element used for dominance analysis.}

\item{sets}{List of vectors of names; each list element's vector is a set.}

\item{all}{Vector of names in all subsets.}
}
\description{
Dominance analysis procedure accepting user-defined models and fit statistics.  An R port of Stata's -domin- module.
}
\examples{
domin()
}
\keyword{importance}
\keyword{relative}
