% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domin.r
\name{domin}
\alias{domin}
\title{Dominance analysis support for \code{formula}-based functions}
\usage{
domin(
  formula_overall,
  reg,
  fitstat,
  sets = NULL,
  all = NULL,
  complete = TRUE,
  ...
)
}
\arguments{
\item{formula_overall}{An object of class \code{\link{formula}} or that can be coerced to class \code{formula} for use in the functon in \code{reg}.}

\item{reg}{The function implementing the predictive(or "reg"ression) model called. Uses \code{\link{do.call}} and accepts any function call \code{do.call} would accept.}

\item{fitstat}{List of specifications to call a fit statistic extracting function (see details). Like \code{reg}, uses \code{do.call} and accepts any function call \code{do.call} would accept.}

\item{sets}{An optional list containing vectors of variable or factor names.  Each \link{vector} in the list is used as a set and always included together in the \code{formula}.}

\item{all}{An optional vector of variable or factor names to be built into the \code{formula} and included in all model subsets.}

\item{complete}{logical.  If \code{FALSE} then complete dominance matrix is not computed.}

\item{...}{additional arguments passed to \code{reg} function.}
}
\value{
Returns an object of \code{\link{class}} "domin"

General_Dominance Vector of general dominance statistics.
}
\description{
Computes dominance statistics for predictive modeling functions that accept a \code{formula}.
}
\details{
The entries in \code{formula_overall}, \code{sets}, and \code{all} are processed
}
\examples{
## Basic linear model with r-square
domin(mpg ~ am + vs + cyl, "lm", list("summary", "r.squared"), data=mtcars)
## Including sets
domin(mpg ~ am + vs + cyl, "lm", list("summary", "r.squared"), data=mtcars, sets=list(c("carb", "gear"), c("disp", "wt")))
## Multivariate linear model with custom multivariate r-square function and all subsets variable
Rxy <- function(obj, names, data) {return(list("r2" = cancor(predict(obj), as.data.frame(mget(names,as.environment(data))))[["cor"]][1]^2))}
domin(cbind(wt, mpg) ~ vs + cyl + am, "lm", list("Rxy", "r2", c("mpg", "wt"), mtcars), data = mtcars, all=c("carb"))
}
\keyword{multivariate}
\keyword{utilities}
