% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance_internals.r
\name{dominance_scalar}
\alias{dominance_scalar}
\title{Internal dominance analysis computation function assuming scalar
output}
\usage{
dominance_scalar(
  fitting_fun,
  args_list,
  cons_args,
  full_fit,
  conditional,
  complete,
  reverse
)
}
\description{
Internal dominance analysis computation function assuming scalar
output
}
\keyword{internal}
