% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domir.r
\name{domir}
\alias{domir}
\alias{domir.formula}
\alias{domir.Formula}
\alias{domir.list}
\title{Dominance analysis methods}
\usage{
domir(.obj, ...)

\method{domir}{formula}(
  .obj,
  .fct,
  .set = NULL,
  .wst = NULL,
  .all = NULL,
  .adj = NULL,
  .cdl = TRUE,
  .cpt = TRUE,
  .rev = FALSE,
  ...
)

\method{domir}{Formula}(...)

\method{domir}{list}(...)
}
\arguments{
\item{.obj}{A \code{\link{formula}}, \code{\link[=formula]{Formula}}, or \code{\link{list}}.

Parsed to produce subsets.  Always submits subsets of the same type
to \code{.fct}.}

\item{...}{Passes arguments to other methods; passes arguments to
function in \code{.fct}.}

\item{.fct}{A function/closure or string function name.

Applied to all subsets received from \code{.obj}.  Must return a length 1
(scalar), numeric vector.}

\item{.set}{A \code{list}.

Must be comprised of elements of the same class as \code{.obj}.  Elements of
the list can be named.}

\item{.wst}{Not yet used.}

\item{.all}{A \code{formula}, \code{Formula}, or \code{list}.

Must be the same class as \code{.obj}.}

\item{.adj}{A \code{formula}, \code{Formula}, or \code{list}.

Must be the same class as \code{.obj}.}

\item{.cdl}{Logical.

If \code{FALSE} then conditional dominance matrix is not computed and
method to produce general dominance statistics changes.}

\item{.cpt}{Logical.

If \code{FALSE} then complete dominance matrix is not computed.}

\item{.rev}{Logical.

If \code{TRUE} then standardized vector, ranks, and complete dominance
designations are reversed in their interpretation.}
}
\value{
Returns an object of \code{\link{class}} "domir" which is a list composed of:

\describe{
\item{\code{General_Dominance}}{Vector of general dominance values.}
\item{\code{Standardized}}{Vector of general dominance values normalized
to sum to 1.}
\item{\code{Ranks}}{Vector of ranks applied to the general dominance values.}
\item{\code{Conditional_Dominance}}{Matrix of conditional dominance values.
Each row represents a name; each column represents a number of names
included in \code{.fct}.}
\item{\code{Complete_Dominance}}{Logical matrix of complete dominance
designations. The term represented in each row indicates dominance status;
the terms represented in each columns indicates dominated-by status.}
\item{\code{Value}}{Value returned by \code{.fct} with all names included.}
\item{\code{Value_All}}{Value of \code{.fct} associated with names included
in \code{.all}.}
\item{\code{Value_Adjust}}{Value of \code{.fct} associated
with names in \code{.adj}.}
\item{\code{Call}}{The matched call.}
}
}
\description{
Parses input depending on method, computes all valid subsets,
submits subsets to a function, and computes dominance decomposition
statistics based on the returned values from the function.
}
\details{
\code{.obj} is used to determine the number of subsets included the
dominance analysis.  How the subsets are determined depends on \code{.obj}s
class method.

The \code{formula} method applies the standard \code{\link{terms}} function parsing
which separates elements on the right hand side of the \code{formula}.  All
elements separated by \code{+} are considered a separate term that contributes to
determining subsets. The \code{Formula} and \code{list} methods are not
yet implemented.

All methods for \code{domir} will submit subsets generated to the
function in \code{.fct} as the first, unnamed argument.

All subset generation methods' default behavior of using all elements
as separate subset factors can be overriden using \code{.set}, \code{.all},
and \code{.adj} arguments.

\code{.set} binds together elements in \code{.obj} such that the form their own term.

\code{.all} includes elements in \code{.obj} all subsets including their contribution
to the result.

\code{.adj} includes elements in \code{.obj} all subsets removing their contribution
to the result.
}
\examples{
## Basic linear model with r-square

lm_r2 <- function(fml, data) { 
  lm_res <- lm(fml, data = data)
  r2 <- summary(lm_res)[["r.squared"]]
  return(r2) }

domir(mpg ~ am + vs + cyl, 
  lm_r2,
  data = mtcars)


## Linear model including set

domir(mpg ~ am + vs + cyl + carb + gear + disp + wt,
 lm_r2,
 .set = list(~ carb + gear, ~ disp + wt),
 data = mtcars)


## Multivariate linear model with custom multivariate r-square function 
## and all subsets variable

mlm_rxy <- function(fml, data, dvnames) {
  mlm_res <- lm(fml, data = data)
  mlm_pred <- predict(mlm_res)
  mlm_rxy <- cancor(mlm_pred, data[dvnames])$cor[[1]]^2
  return(mlm_rxy)
  }
       
domir(cbind(wt, mpg) ~ vs + cyl + am + carb,
  mlm_rxy, 
  .all = ~ carb,
  data = mtcars, 
  dvnames = c("wt", "mpg"))


## Named sets

domir(mpg ~ am + gear + cyl + vs + qsec + drat,
  lm_r2,
  data = mtcars, 
  .set = list(trns = ~ am + gear, 
    eng = ~ cyl + vs, misc = ~ qsec + drat))
  
## Linear model using AIC

lm_aic <- function(fml, data) { 
  lm_res <- lm(fml, data = data)
  aic <- AIC(lm_res)
  return(aic) }

domir(mpg ~ am + carb + cyl, 
  lm_aic, 
  .adj = ~ 1,
  .rev = TRUE,
  data = mtcars)
}
