% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list.R
\name{fmllst2Fml}
\alias{fmllst2Fml}
\title{Translate \code{formula_list} into \code{Formula::Formula}}
\usage{
fmllst2Fml(fmllst, drop_lhs = NULL)
}
\arguments{
\item{fmllst}{A \code{formula_list} classed object.}

\item{drop_lhs}{An integer vector.

Used as a selection vector to remove left hand side names prior to
generating the \code{Formula} object. This vector must be composed of
integers (e.g., 1L and not 1).}
}
\value{
A \code{Formula::Formula} object.
}
\description{
Translates \code{{domir}}s \code{formula_list} objects into an a \code{\link[Formula:Formula]{Formula::Formula}}
}
