% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list.R
\name{fmllst2Fml}
\alias{fmllst2Fml}
\title{ATranslate \code{formula_list} into \code{Formula::Formula}}
\usage{
fmllst2Fml(fmllst)
}
\arguments{
\item{fmllst}{A \code{formula_list} classed object.}
}
\value{
A \code{Formula::Formula} object.
}
\description{
Translates \code{{domir}}s \code{formula_list} objects into an a \code{\link[Formula:Formula]{Formula::Formula}}
}
