% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list.R
\name{formula_list}
\alias{formula_list}
\title{A \code{\link{list}} composed of \code{formulas}}
\usage{
formula_list(...)
}
\arguments{
\item{...}{\code{formula}s, possibly named}
}
\value{
A \code{list} of class \code{formula_list}.
}
\description{
Defines a list object class composed of \code{formula} objects that
is used to obtain RHS-LHS pairs in its \code{\link{domir}} S3 method.
}
\details{
All \code{formula_list}s enforce requirements that the list are composed of
individual \code{formula}s and that each formula is unique with its own,
different, non-\code{NULL} dependent variable/response.
}
