% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domin.r
\name{print.domin}
\alias{print.domin}
\title{Print method for \code{domin}}
\usage{
\method{print}{domin}(domin_obj)
}
\description{
Reports basic results from \code{domin} run
}
\keyword{importance}
\keyword{relative}
