% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domin.r
\name{print.domin}
\alias{print.domin}
\title{Print method for \code{domin}}
\usage{
\method{print}{domin}(domin_obj)
}
\description{
Dominance analysis procedure accepting user-defined models and fit statistics.  An R port of Stata's -domin- module.
}
\keyword{importance}
\keyword{relative}
