% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domir.r
\name{print.domir}
\alias{print.domir}
\title{Print method for \code{domir}}
\usage{
\method{print}{domir}(x, ...)
}
\arguments{
\item{x}{an object of class "domir".}

\item{...}{further arguments passed to \code{\link{print.default}}.}
}
\value{
The submitted "domir" object, invisibly.
}
\description{
Reports formatted results from \code{domir} class object.
}
\details{
The print method for class \code{domir} objects reports out the
following results:
\itemize{
\item{Value when all names are included in \code{.fct}.  The value for the
names included in all subsets is reported here if there are any entries
in \code{.all}.  The value for names/terms used to adjust the values is
reported here if there are any entries in \code{.adj}.}
\item{Matrix describing general dominance values, standardized
general dominance values, and the ranking of the general
dominance values}
\item{If \code{.cdl} is \code{TRUE}, matrix describing the conditional
dominance values}
\item{If \code{.cpt} is \code{TRUE}, matrix describing the complete
dominance designations}
\item{If following \code{summary.domir}, matrix describing the strongest
dominance designations between all names}}
The \code{domir} print method alters dimension names for readability and they
do not display as stored in the original \code{domir} object.
}
